Removed lots of redundant code.
