+++ raw_bsd.go
@@ -16,15 +16,6 @@ import (
 	"golang.org/x/net/bpf"
 )
 
-const (
-	// bpfDIn tells BPF to pass through only incoming packets, so we do not
-	// receive the packets we send using BPF.
-	bpfDIn = 0
-
-	// osFreeBSD is the GOOS name for FreeBSD.
-	osFreeBSD = "freebsd"
-)
-
 // bpfLen returns the length of the BPF header prepended to each incoming ethernet
 // frame.  FreeBSD uses a slightly modified header from other BSD variants.
 func bpfLen() int {
@@ -35,7 +26,7 @@ func bpfLen() int {
 		bpfXHeaderLen = 26
 	)
 
-	if runtime.GOOS == osFreeBSD {
+	if runtime.GOOS == "freebsd" {
 		return bpfXHeaderLen
 	}
 
@@ -278,25 +269,6 @@ func configureBPF(fd int, ifi *net.Interface, proto uint16) (int, error) {
 	}
 
 	return buflen, nil
-}
-
-// setBPFDirection enables filtering traffic traveling in a specific direction
-// using BPF, so that traffic sent by this package is not captured when reading
-// using this package.
-func setBPFDirection(fd int, direction int) error {
-	_, _, err := syscall.Syscall(
-		syscall.SYS_IOCTL,
-		uintptr(fd),
-		// Even though BIOCSDIRECTION is preferred on FreeBSD, BIOCSSEESENT continues
-		// to work, and is required for other BSD platforms
-		syscall.BIOCSSEESENT,
-		uintptr(unsafe.Pointer(&direction)),
-	)
-	if err != 0 {
-		return syscall.Errno(err)
-	}
-
-	return nil
 }
 
 // assembleBpfInsn assembles a slice of bpf.RawInstructions to the format required by
